pyproject.toml
[tool.poetry]
name = "poetry-demo-01"
version = "0.1.0"
description = ""
authors = ["nguyen hung <hung.nguyen@tnicorporation.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
flask = "^3.0.3"
firebase-admin = "^6.5.0"
flask-sqlalchemy = "^3.1.1"
flask-jwt-extended = "^4.6.0"
pyjwt = "^2.9.0"
mysqlclient = "^2.2.4"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
flask-restplus = "^0.13.0"
flask-swagger-ui = "^4.11.1"
flask-cors = "^5.0.0"
flask-restful = "^0.3.10"
gunicorn = "^23.0.0"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
